Trying to deploy to Azure

+++ public/js/script.js
@@ -191,6 +191,8 @@ function removeSupportingTables(removalTableData)
 	var i = 0;
 	for (rmData of removalTableData)
 	{
+		if (i == 0) continue;
+		i++;
 		websocket.dropTable(rmData["Material Cost"]);
 		websocket.dropTable(rmData["Subcontractor Fee"]);
 	}
